The basis of the website
These files form the backbone of the code structure. The code written so far is mostly pointers to each of the files with some minor styles and elements added for testing purposes. At this point the code runs but is not pretty. Lots of work still needs to be done to get to the finished product. 

This code will eventually be loaded to letschill.ca for the Let's Chill music duo. The code written here is pro bono and not intended to be sold. If you do use snipped of this code please reference the author Kevin Barker in a comment next to the code. I'm obviously not going to know if you do it or not so I'm trusting whoever is reading this to do the right thing!
Note you will need to the BootstrapStyle.css file if you are going to run this code with the intended styles.