Implemented Font-Awesome icons
To implement Font Awesome Icons i needed to change the location of my
library folder to inside my project folder. This is to prevent firefox's
security from freaking out over the location of the assets.

Also added a few more elements to the scss file that allowes for content
to be centetered within a div element. These classes are the .outer,
.middle and .inner classes.

Implemented better navbar styling with the new font awesome icons,
including hover animations and centering on the collapsed navbar.