Implementing Video scroll
I talked with the producers and we brainstormed the idea of having a
video that plays as you scroll. I started to implement this feature with
a webm file i created. Still lots of work to be done but the base
functionality is there.

TO DO:
implement smoother scroll listener, very choppy right now
Fix video css to replace the fixed record image currently in the
background
Implement horizontal video for mobile devices when they are in portrait
mode, using media queries.