Dropped scrolling video
Scrolling video was so close to being finished but the final push was
killing me. I was struggling to implement the idea because even though
the code was working the video playback was very choppy. I'm at a bit of
a loss as to why because the test video I used was larger than the video
I actually needed and it worked perfectly so it wasn't a performance
issue. I think if I played with differet codects and resolutions and
framerates I could have got it to work, but the truth is while the idea
was kind of neat it really did slow the site down and took away from the
user experience.
So i talked to my clients and we decided to cut it and just have a
looping video instead. This was great because i could up the file size
again to the proper resolution which honestly just looked better
overall.
There is still some things I need to work out, currently getting the
video to run on mobile devices is being kind of annoying so I will still
need to play around with codects and framerates there (oh good) but at
least this time there is way more support online for getting videos to
run on mobile.
I also removed quite a bit of files that aren't being used anymore,
these elements were more placeholders anyway until the final vision came
through, and as we are closing in on the final vision these assets are
not needed anymore.